Fix Gibbs solder, fix Java Gaussian factor functions, add some factor functions
Fix Gibbs solver, which entails changing the factor-function interface to add evalEnergy, to get the factor value in the log domain rather than linear domain to avoid numerical precision issues.  Modified most Java factors to do the same.  Future factors should also do this, and eventually we should phase out overriding eval directly.

Also fixed a bug in Gaussian factor functions, which were missing the factor of 2 in the exponent (so the variance was wrong).

Also, added some factor-functions.  The primary one being a parameterized Gaussian factor, for which the mean and inverse-variance can be variables so that they can be inferred from the data.  Also added a few other possibly useful factors.