HMM parameter estimate using Gibbs solver
Implemented HMM transition-matrix parameter estimation using the Gibbs solver.  Added two FactorFunctions to support this.  One representing a parameterized HMM transition matrix itself.  And one representing the prior on the transition matrix variables.  The matrix is represented in the -log domain, and ignoring normalization.  As a result of ignoring normalization, the conjugate prior for the matrix elements are independent and Gamma distributed (with constant beta parameter).  In the -log domain, the prior is the NegativeExpGamma.

Also added a test for this functionality.