Speed-up factor creation
AddFactor (and vectorized version) when checking if variable is already in the graph, was asking for all sub-graphs.  This was done by looping through all factors and seeing if it was a sub-graph, which is really slow and grows as the graph grows.  This was a major portion of the addFactor time.  Instead, FactorGraph now keeps a list of subgraphs, and just returns that directly.