Added methods to define and use groups of variables
Added methods to define groups of variables that are not necessarily of the same kind, so that MATLAB can access them with a single Java call without the need for slow MATLAB loops.  The ability to use these groups has been implemented in the Gibbs solver, allowing the ability to set and hold sample values for groups of variables.  The group mechanism itself is general and could be used for other purposes.