Generalize finding index from discrete domain value
Centralized the code to find the index into the elements of a discrete domain from the value (now a method in DiscreteDomain class).  Also, generalized it to accommodate domains that are arrays of doubles or arrays of objects (there may still be other domain types we would like to include as well).