Distinguish Equality factor (undirected) from Equals (and NotEquals) factors that correspond to directed functions, like GreaterThan and LessThan
Equality factor (same as old EqualsDelta, which is still there for backward compatibility).  Equals and NotEquals factors have a directed output indicator Bit variable, like the GreaterThan and LessThan factors.  Also added helper functions in MATLAB to use Equals and NotEquals in lieu of overloading the == and ~= operators, which are needed for comparing variables.  Added test and documentation.