Fix bug in case of mixed real/discrete variable graph
For graph with mixed Real and Discrete, fg.Variables and fg.plot crash.  When there are real and discrete variables, these two methods call obj.VectorObject.getVariableVector, which is in PFactorGraphVector.java, which calls PHelpers.convertToVariableVector, which creates a generic PVariableVector.  Later when wrapProxyObject.m is called, it checks the type, which is none-of-the-above, which falls through the else case.  This assumes it's a Real.  But if some of the variables are Discrete, this isn't the correct thing to do.