Fix Supers.invokeMethod problem with IllegalAccessException on non-public instance
Previously if you tried to invoke a public method that was originally
declared in a public superclass or interface through an object whose
runtime type is not public, you would get an IllegalAccessException.
This fix works around that by looking up the method on a public superclass
or interface, if there is one.